Move interfaces to their own module
