+++ src/interfaces/TileAsset.ts
@@ -1,0 +1,5 @@
+import { TileAsset } from '../types';
+
+export interface TileAssetI {
+  create: (image: string) => TileAsset;
+}
