+++ src/interfaces/index.ts
@@ -1,0 +1,3 @@
+export * from './Deck';
+export * from './Tile';
+export * from './TileAsset';
