+++ src/types/Deck.ts
@@ -1,20 +1,7 @@
-import { BaseSyntheticEvent } from 'react';
-import { TileI, Tile as TileT } from './Tile';
-import { TileAsset as TileAssetT } from './TileAsset';
+import { Tile as TileT } from './Tile';
 
 export type Deck = {
   tiles: TileT[];
   selectedTiles: TileT[];
   afterEffect?: Deck | null;
 };
-
-export interface DeckI {
-  init: (tiles: TileAssetT[]) => Deck;
-  findTile: (deck: Deck, tileId: string) => TileT | undefined;
-  processTile: (
-    e: BaseSyntheticEvent,
-    tile: TileT,
-    deck: Deck,
-    tileAPI: TileI,
-  ) => Promise<Deck>;
-}
